+++ CHANGELOG.md
@@ -1,5 +1,28 @@
-## [TBD]:
-* Increased macOS minimum version to 10.15 #2220
+## [1.6.3]
+* Merge 1.6.1-hotfix
+
+## [1.6.2]:
+* Add native auth instructions to error description when reset password required is returned (#2582)
+* Save error received from ESTS, and return it to the client on silent broker calls (#2379)
+* Support web_page_uri (#2384)
+
+## [1.6.1-hotfix]
+* Add support of "lookup" mode in broker #2414
+
+## [1.6.1]:
+* Support extra query parameters on logout endpoint (#2339)
+* Add support functions to help broker improve cross cloud experience (#2361)
+
+## [1.6.0]:
+* Support extra query parameters on logout endpoint (#2339)
+
+## [1.5.1]:
+* Parse and add STS error codes in token error result (#2319)
+* VisionOS support added (#2139)
+
+## [1.5.0]
+* Increased macOS minimum version to 10.15 (#2220)
+* Added Native Auth feature for MacOS to interact with the Microsoft Entra ID services (#2220)
 
 ## [1.4.2]
 * Fix invalid cert auth request handling (hotfix). (#2278)